// web/src/app/users-tasks/tasks-services/obszar.enum.ts
export enum Obszar {
    W_PIERWSZEJ_CHWILI = "W pierwszej chwili",
    MATERIALY_REFERENCYJNE = "Materiały referencyjne",
    OBOWIAZKI = "Obowiązki",
    W_NIEDALEKIEJ_PRZYSZLOSCI = "W niedalekiej przyszłości",
    MOZE_KIEDYS = "Może kiedyś",
    UKONCZONE = "Zadania ukończone" //==kosz
}